avezov/avezov_university/requirements.txt
Django>=4.2,<6.0
sqlparse==0.5.4
tzdata==2025.3
django-jazzmin
python-dotenv
aiohttp
Pillow
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-jazzmin
django_ckeditor_5
drf_yasg
django_filter
requests
python-telegram-bot==20.8
python-dotenv
celery
redis
social-auth-app-django
